Job scheduling queues (#42)
Sometimes threading can be tricky. 
it's easy to forget to lock a critical section, and cause thread safety issue. 
it's also easy to mistakenly lock a non-critical section, and cause performance issue.  

Instead of having `synchronized` blocks everywhere in the `MvRxStateStore`, it can be helpful to encapsulate the dangerous parts and prevent future mistakes

@BenSchwab @gpeal @elihart