Use tuples for selectSubscribe (#45)
Migrates selectSubscribe to use a tuple with distinctUntilChanged. This is a silent API change that will now immediately deliver the initial value. The updated tests reflect this new behavior.

This also works without scan or ever looking at oldState which is great.

This supersedes #41 

Thanks @hellohuanlin and @BenSchwab for raising some good discussions that led to this

@BenSchwab @hellohuanlin