Removed lifecycleAwareLazy value restriction (#55)
We tried restricting accessing the lifecycleAwareLazy before CREATED. However, the lifecycle is still initialized in a Fragment's onCreate which makes this impractical. It was merely a safeguard before so we can remove it.

Also fixed another test that has a new error class because the lifecycleAwareLazy initializes it immediately and creates an InvocationException which bubbles up to a RuntimeException.