Improve impure error messages (#160)
Previously, MvRx would just throw an exception that said "Your reducer is not pure!".
However, they will now say:
Your reducer is not pure! count changed from 1 to Ensure that your state properties properly implement hashCode. Note that it says the actual property that changed and what it changed to.

It will also highlight if you forgot data on one of your Kotlin classes because the toString() will not be formatted.