Small cleanup to get rid of Kotlin nullability warning (#187)
Cleanup restoreViewModels to get rid of Kotlin nullity warning.

Here was the warning. This happens because getParcelable returns Nullable.

w: MvRxViewModelStore.kt: (107, 68): Type mismatch: inferred type is MvRxPersistedViewModelHolder? but MvRxPersistedViewModelHolder was expected
w: MvRxViewModelStore.kt: (108, 67): Type mismatch: inferred type is MvRxPersistedViewModelHolder? but MvRxPersistedViewModelHolder was expected
Additionally, used when statement instead of optional casting to Fragment and then force casting to FragmentActivty if casting to Fragment was not successful. Now, the type checking is more clear. Everything is smart-casted.