Upgrade Kotlin to 1.3.50, AGP to 3.5.0, and Gradle to 5.6 (#275)
This seems to fix #256.
However, it didn't crash while I was developing it, started crashing somehow and even after upgrading, it crashed until I moved the function and moved it back and now it isn't crashing. I tried various combinations of running clean and rm -rf */build/**.

Upgrading these things should be good anyway and hopefully it also fixes this.

I also:

Upgraded a few other dependencies such as AGP but the that broke robolectric so I upgraded that and removed shadows-support-v4 since jetifier now breaks on it and we don't need it anymore.
Fixed a new lint error that was caught after upgrading (ArrayMap sdk check)
Fixed detekt errors since I have them as a pre-push hook.