Replaced internal RxJava with Coroutines (#347)
This PR converts the internal implementation from RxJava to Coroutines. This was not designed to be a breaking change so no external APIs still use Disposable.

However, a new API was added: ViewModel.stateFlow This property allows consumers to easily combine MvRx state with other operations or chain it to other Flow operators. Thoughts on this API? It doesn't violate the primary concern of providing synchronous state access from within a ViewModel.

Once this is landed, I think it is time to create the mvrx-rxjava2 artifact and start splitting off RxJava for good. Thoughts?

Fixes #296
Fixes #79
#130