[2.0] Removed the need for a base Fragment class (#396)
I was able to get the core MvRx functionality working with just the MvRxView interface, no base class needed! This has been one of the most requested features and is a blocker for some people to use MvRx entirely.

#382 should no longer be necessary after this.

##Breaking Change##
Since MvRxView is just an interface, if no ViewModels are used and no MvRx code is run, we can't call postInvalidate() to cause invalidate to be called. Simply using a view model delegate is enough to trigger that behavior, however. I think this breaking change is worth it and I made a note of it in the changelog.

@denis-bezrukov