Make BaseMvRxViewModel not extend ViewModel (#394)
It would be amazing if we can get BaseMvRxViewModel to be a pojo. That would enable testing on the jvm without robolectric and a potential path to Kotlin Multiplatform.

I was just poking around at what this would look like and it turned out to be fairly simple.

The next hurdle is replacing the inter-vm subscriptions which use Lifecycle with something else but that should be doable.