Added docsify docs (#411)
This is the first round of updating docs for Mavericks 2.0. I used docsify which I use for Lottie here. We should be able to get airbnb.io/mvrx to point to these.
Docsify is nice because you just write markdown and it generates the page locally so no html generation is necessary.

About half of the docs are new/rewritten and half are copied over from the wiki and lightly edited.

You can try this out locally with npm i -g docsify then docsify serve ./docs

This is just a first pass. We'll need to clean these up and add more before release. Any help making the documentation more robust would be greatly appreciated.