Improve performance of lifecycle aware flow (#444)
we noticed a performance regression when main thread is busy
after reviewing combineTransform implementation I found that it converts flows to fair channels. Fair channel uses yield each time in fast-path offer, that means it dispatches coroutine per each element.

In this PR I rewrote flowWhenStarted by forking combineTransform implementation and adopting it for us. It includes two changes:

use non-fair channels
end flow if startedFlow/vm.stateflow is ended (combineTransform waited for completion of both flows, so we used this.onCompletion { startedChannel.cancel() })
I also want to cover this with more tests for
a) prevent this regression in future
b) ensure nothing was broken due change to non-fair channel
But operator is ready for review.