[Compose] Create mavericks-compose (#514)
This PR introduces official Mavericks Compose APIs.
Usage is as follows:

@Composable
fun MyComponent() {
    val viewModel: CounterViewModel = mavericksViewModel()
    val count by viewModel.collectState(CounterState::count)
    // or
    val state by viewModel.collectAsState()
    Button(onClick = activityScopedViewModel::incrementCount) {
        Text("Count: $count")
    }
}
Things seem to be working well and you can scope the view model to arbitrary places by overriding the lifecycle owner (and optionally the ViewModelStoreOwner and SavedStateRegistryOwner).

Ideas and other thoughts welcome but this seems to be a nice parallel to the Jetpack ViewModel interop.

There were a few small changes needed to base Mavericks so I'll put those up in separate individual PRs as well.

For this PR, we could:

Land this on master. This would mean that master is using AGP 7 alphas which should be fine but isn't ideal.
Land this on a long running feature branch jetpack-compose and merge future PRs and changes on master into that branch.
I think just going straight for option 1 is fine but I'm open to a discussion. The only downside is that we'll have to use Arctic Fox when working on Mavericks and will also need to keep the AGP version in sync with the latest canary versions but we'll have to do that for the compose branch anyway.

I'm also open to ideas on versioning and release cadence.

Fixes #437