Add covariant recreation support (#565)
* Add covariant recreation support

Adds support for recreating MavericksViewModels and MavericksState from a persisted Bundle when the declaration site uses a covariant type. This is accomplished by serializing the Java Class of both the ViewModel and State types, and storing them in the StateRestorer class. These values are then used in MavericksFactory and MavericksStateFactory if present, overriding the declaration site types.

Also adds a test in StateRestorationTest.kt to cover this newly supported behavior. Without these changes, the test fails with an IllegalStateException when trying to instantiate the abstract State class from the declaration site.

* Fix typo in error message

Co-authored-by: Ben Schwab <bschwab@coursera.org>

Co-authored-by: Allen Chen <allen.chen@airbnb.com>
Co-authored-by: Ben Schwab <bschwab@coursera.org>