Handle repeated state restorations where declaration of ViewModel does not use original types
Covers an edge case from https://github.com/airbnb/mavericks/pull/566 - if we are saving state when a ViewModel is first created, we need to use the classes from the StateRestorer if it is present. In the case where we are restoring from a previously restored state, only the restored state has knowledge of the original declaration site classes that we need to call.